Refactoring, fix wrong input (angle vs vector) - Distributions changed (unified) - Fixed inf critical energy - Speedups, old algorithm supports backscattering - Fixed bandwidth correction - spectrum: natural log to log10 - Removed desorption export (deprecated)

+++ TexturePlotter.h
@@ -57,7 +57,7 @@ private:
   Facet        *selFacet;
   float        lastUpdate;
   float			maxValue;
-  int			maxX,maxY;
+  size_t			maxX,maxY;
   char         currentDir[512];
 
   GLList      *mapList;
